+++ app/traits/trait-empty-init.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * Empty init trait.
+ *
+ * @link https://vcore.au
+ *
+ * @package CF_Images
+ * @subpackage CF_Images/App/Traits
+ * @author Anton Vanyukov <a.vanyukov@vcore.ru>
+ * @since 1.8.1
+ */
+
+namespace CF_Images\App\Traits;
+
+if ( ! defined( 'WPINC' ) ) {
+	die;
+}
+
+/**
+ * Empty_Init trait.
+ *
+ * @since 1.8.1
+ */
+trait Empty_Init {
+	/**
+	 * Init module.
+	 *
+	 * @since 1.8.1
+	 */
+	public function init() {}
+}