+++ assets/_src/context/provider.tsx
@@ -17,8 +17,14 @@ import SettingsContext from './settings';
  * @class
  */
 const SettingsProvider = ({ children }: { children: ReactElement[] }) => {
-	const { cfStatus, fuzion, hideSidebar, isNetworkAdmin, settings } =
-		window.CFImages;
+	const {
+		browserTTL,
+		cfStatus,
+		fuzion,
+		hideSidebar,
+		isNetworkAdmin,
+		settings,
+	} = window.CFImages;
 
 	const [stats, setStats] = useState(window.CFImages.stats);
 	const [modules, setModules] = useState(settings);
@@ -45,6 +51,7 @@ const SettingsProvider = ({ children }: { children: ReactElement[] }) => {
 	return (
 		<SettingsContext.Provider
 			value={{
+				browserTTL,
 				modules,
 				setModule,
 				noticeHidden,